modified config variables

+++ lib/randomNumber.go
@@ -10,7 +10,7 @@ import (
 // To generate a random number consisting of x number of digits, pass x as
 // the parameter. For example, SecureRandomNumber(3) will generate a number
 // between 100 and 999
-func SecureRandomNumber(totalDigit uint) uint64 {
+func SecureRandomNumber(totalDigit uint64) uint64 {
 	if totalDigit == 0 {
 		return 0
 	}
